Make reflector.ini.php & scanner.inc.php PSR-4 complient
