+++ src/PHPTracerWeaver/Reflector/ClassCollatorInterface.php
@@ -1,0 +1,6 @@
+<?php namespace PHPTracerWeaver\Reflector;
+
+interface ClassCollatorInterface
+{
+    public function collate($first, $second);
+}
