+++ src/PHPTracerWeaver/Reflector/DummyClassCollator.php
@@ -1,0 +1,9 @@
+<?php namespace PHPTracerWeaver\Reflector;
+
+class DummyClassCollator implements ClassCollatorInterface
+{
+    public function collate($first, $second)
+    {
+        return 'mixed';
+    }
+}
