+++ src/PHPTracerWeaver/Scanner/FileAccessInterface.php
@@ -1,0 +1,9 @@
+<?php namespace PHPTracerWeaver\Scanner;
+
+/** provides access to a file */
+interface FileAccessInterface
+{
+    public function getContents();
+
+    public function getPathname();
+}
