+++ src/PHPTracerWeaver/Scanner/ScannerInterface.php
@@ -1,0 +1,7 @@
+<?php namespace PHPTracerWeaver\Scanner;
+
+/** a statemachine for scanning a tokenstream  */
+interface ScannerInterface
+{
+    public function accept(Token $token);
+}
